+++ hurocon/cli/__init__.py
@@ -1,7 +1,7 @@
 def run_cli():
-    from .root import cli
+    from .__cli_base__ import cli
 
     # Initialize CLI modules
-    from . import auth, config, device, sms, lte
+    from . import auth, config, device, sms, lte, net
 
     cli()