+++ hurocon/cli/auth.py
@@ -4,7 +4,7 @@ import click
 from click_didyoumean import DYMGroup
 from huawei_lte_api.Client import Client
 
-from .root import cli
+from .__cli_base__ import cli
 from ..core.local_cfg import AuthConfig
 from ..core.connection import HRC_Connection
 from ..core.const import LOCAL_CONFIG_DEFAULT