+++ hurocon/meta.py
@@ -1,2 +1,2 @@
-version = '0.4.5'
+version = '0.5.0dev0'
 author = 'maximilionus'